requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
sqlalchemy_utils
python-dotenv
werkzeug
fastapi_jwt_auth